Change CFX_DIBitmap::MultiplyAlpha() to take a float
Some MultiplyAlpha() methods take a float, while others take an integer.
It is too easy to mix them up, so change CFX_DIBitmap::MultiplyAlpha()
to take a float to avoid inconsistencies.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/115452
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ core/fxge/dib/cfx_dibitmap.h
@@ -65,7 +65,7 @@ class CFX_DIBitmap final : public CFX_DIBBase {
   bool SetUniformOpaqueAlpha();
 
   // TODO(crbug.com/pdfium/2007): Migrate callers to `CFX_RenderDevice`.
-  bool MultiplyAlpha(int alpha);
+  bool MultiplyAlpha(float alpha);
   bool MultiplyAlphaMask(const RetainPtr<const CFX_DIBBase>& pSrcBitmap);
 
   bool TransferBitmap(int dest_left,
